Added terminal commands to set/get content
Currently changes delta time variable temporarily. Resets on boot but
will be able to use the terminal command for debugging such as seeing
data flow for the SD card reader to check that it is working properly.
The watchdog is also beginning to be added so it can be a reliable
console system